Move Context & SelectBase to abc
