+++ docs/api_reference/select.md
@@ -5,14 +5,4 @@ description: Select API reference
 
 # Select API Reference
 
-::: miru.select.base
-
-::: miru.select.text
-
-::: miru.select.user
-
-::: miru.select.role
-
-::: miru.select.channel
-
-::: miru.select.mentionable
+::: miru.select
