+++ docs/guides/error_handling.md
@@ -18,68 +18,34 @@ it will, by default, print the exception and traceback to the console.
 
 ## Example:
 
-=== "Gateway"
+```py
+import hikari
+import miru
 
-    ```py
-    import hikari
-    import miru
+class ErrorView(miru.View):
 
-    class ErrorView(miru.View):
+    @miru.button(label="Click me!", style=hikari.ButtonStyle.SUCCESS)
+    async def basic_button(self, ctx: miru.ViewContext, button: miru.Button) -> None:
+        await ctx.respond("You clicked me!")
 
-        @miru.button(label="Click me!", style=hikari.ButtonStyle.SUCCESS)
-        async def basic_button(self, button: miru.Button, ctx: miru.ViewContext) -> None:
-            await ctx.respond("You clicked me!")
+    # Create a button that raises an exception
+    @miru.button(label="Error", style=hikari.ButtonStyle.DANGER)
+    async def error_button(self, ctx: miru.ViewContext, button: miru.Button) -> None:
+        raise RuntimeError("I'm an error!")
 
-        # Create a button that raises an exception
-        @miru.button(label="Error", style=hikari.ButtonStyle.DANGER)
-        async def error_button(
-            self, button: miru.Button, ctx: miru.ViewContext
-        ) -> None:
-            raise RuntimeError("I'm an error!")
+    # Define our custom error-handler
+    # Keep in mind that if you override the error-handler,
+    # tracebacks will no longer be printed in the console,
+    # unless you call super's implementation.
+    async def on_error(
+        self,
+        error: Exception,
+        item: miru.ViewItem | None = None,
+        ctx: miru.ViewContext | None = None
+    ) -> None:
+        # ctx is only passed if the error is raised in an item callback
+        if ctx is not None:
+            await ctx.respond(f"Oh no! This error occured: {error}")
+```
 
-        # Define our custom error-handler
-        # Keep in mind that if you override the error-handler,
-        # tracebacks will no longer be printed in the console,
-        # unless you call super's implementation.
-        async def on_error(
-            self,
-            error: Exception,
-            item: miru.ViewItem | None = None,
-            ctx: miru.ViewContext | None = None
-        ) -> None:
-            # ctx is only passed if the error is raised in an item callback
-            if ctx is not None:
-                await ctx.respond(f"Oh no! This error occured: {error}")
-    ```
 
-=== "REST"
-
-    ```py
-    import hikari
-    import miru
-
-    class ErrorView(miru.View):
-
-        @miru.button(label="Click me!", style=hikari.ButtonStyle.SUCCESS)
-        async def basic_button(self, button: miru.Button, ctx: miru.ViewContext) -> None:
-            await ctx.respond("You clicked me!")
-
-        # Create a button that raises an exception
-        @miru.button(label="Error", style=hikari.ButtonStyle.DANGER)
-        async def error_button(self, button: miru.Button, ctx: miru.ViewContext) -> None:
-            raise RuntimeError("I'm an error!")
-
-        # Define our custom error-handler
-        # Keep in mind that if you override the error-handler,
-        # tracebacks will no longer be printed in the console,
-        # unless you call super's implementation.
-        async def on_error(
-            self,
-            error: Exception,
-            item: miru.ViewItem | None = None,
-            ctx: miru.ViewContext | None = None
-        ) -> None:
-            # ctx is only passed if the error is raised in an item callback
-            if ctx is not None:
-                await ctx.respond(f"Oh no! This error occured: {error}")
-    ```
