+++ miru/internal/types.py
@@ -5,16 +5,15 @@ import typing as t
 if t.TYPE_CHECKING:
     import hikari
 
+    from miru.abc.context import Context
     from miru.abc.item import Item, ViewItem
     from miru.abc.item_handler import ItemHandler
-    from miru.context import Context, ViewContext
     from miru.view import View
 
 AppT = t.TypeVar("AppT", bound="hikari.RESTAware")
 BuilderT = t.TypeVar("BuilderT", bound="hikari.api.ComponentBuilder")
 ViewT = t.TypeVar("ViewT", bound="View")
 ViewItemT = t.TypeVar("ViewItemT", bound="ViewItem")
-ViewContextT = t.TypeVar("ViewContextT", bound="ViewContext")
 HandlerT = t.TypeVar("HandlerT", bound="ItemHandler[t.Any, t.Any, t.Any, t.Any, t.Any]")
 ContextT = t.TypeVar("ContextT", bound="Context[t.Any]")
 ItemT = t.TypeVar("ItemT", bound="Item[t.Any, t.Any, t.Any]")
